iommu/vt-d: Tack device aux-attach with device_aux_domain_info.
In existing code, if wanting to loop all devices attached to a domain,
current code can only loop the devices which are full-attached to the
domain, for sub-devices, they are not tracked in domain. This patch adds
struct device_aux_domain_info which is created per domain-device aux-attch.
So that sub-devices are tracked in domain.

This was found by the failure of iommu_domain_get_attr() when passthru mdev
to guest. There is no device in domain->devices, thus unable to get the cap
and ecap of iommu unit. But this domain actually has one sub-device which is
attached via aux-manner. This patch fixes the issue.

Signed-off-by: Liu Yi L <yi.l.liu@intel.com>