iommu: Add /dev/usva
Today, more and more platform vendors support Shared Virtual Addressing(SVA).
Meanwhile, platform vendors also offer nested translation (a.k.a dual-stage
translation). Under such translation mode, there are two stage translations,
in which the most common case is GVA->GPA and GPA->HPA translation. For
userspace applications with a device or subdevice assigned, it needs to have
a way to configure stage-1 translation structure to the IOMMU driver.

This patch comes after a series of discussions on guest SVA uAPIs[1]. A new
/dev/usva interface is introduced to facilitate all use cases related to
nested translation.

This patch comes after a bunch of discussions in the mailing list[1]. It's
going to expose a /dev/usva node to userspace. Open of this file will create
an SVA context for configuring stage-1 translation structure and return a
FD. vSVA is the primary usage of this file today. It is only activated after
associated with a device or subdevice and an iommu domain defines stage-2
translation.

Take vSVA as an example, userspace should ensure the nested translation is
supported and enabled before associating its assigned device with an SVA FD.
If not, the association will be failed. After successful association, userspace
also needs to check the SVA info before heading to configure vSVA usage. e.g.
bind guest page table to host or bind guest PASID table to host.

[1] https://lore.kernel.org/kvm/20210302171551.GK4247@nvidia.com/

Signed-off-by: Liu Yi L <yi.l.liu@intel.com>