vfio: associate vfio device with SVA FD
With the introduction of /dev/usva, userspace application with assigned
devices can use ioctls with SVA FD to configure stage-1 translation
structure to host IOMMU. e.g. guest page table with a specified PASID
for vSVA usages.

In native, SVA operation is per-device. When comes to virtualization, we
also want to keep such philosophy. so this patch associates SVA FD with
a vfio device, thus VM are free to bind stage-1 translation for the assigned
devices.

Signed-off-by: Liu Yi L <yi.l.liu@intel.com>